Added a stub of requirements.txt dep management.
It doesn't do much yet, since I'll likely need to refactor the Project
class to be a more effective tool, and implement config parsing from the
local tutconfig on all other commands through that interface. But some
of that has been stubbed as well, then all that's left to a basic happy
path implementation for one usecase is setting up a test tool.