Tut is now capable of sel hosting.
The big problem with this is that because it doesn't correctly install
itself to the system path yet, that isn't super useful. That is, trying
to actually self-host at this point would require running some
frankenstinian command like `pipenv run 'python tut install requests'`,
which is more trouble than its worth. I'm also not quite sure how the
different pipenvs would play with each other, but I'm not really willing
to go down that path yet.